Add a test for Library.
It doesn't pass, because of other mechanical issues.
For example, shuffling a player's deck currently causes them to throw
their hand into the pile as well (???). This is very wrong, but the
shuffle function is used often enough that I don't want to mess with it
as part of the Library change. Also, of course the current "cards are
discarded as soon as they are played" behavior messes up Library
beyond belief. But once those issues are solved, I think Library is ready.